combined storage and handler ammonia emissions calculator
