+++ tests/test_biophysical/test_manure/test_storage/test_storage.py
@@ -1,5 +1,4 @@
 import pytest
-from pytest_mock import MockerFixture
 
 from RUFAS.biophysical.manure.storage.storage import Storage
 from RUFAS.biophysical.manure.storage.storage_cover import StorageCover
@@ -30,8 +29,7 @@ def test_receive_manure() -> None:
 
 
 @pytest.mark.parametrize(
-    "vol_sols,temp,degradable,expected",
-    [(100.0, 20.0, False, 0.003138872), (100.0, -10.0, True, 0.007100907)]
+    "vol_sols,temp,degradable,expected", [(100.0, 20.0, False, 0.003138872), (100.0, -10.0, True, 0.007100907)]
 )
 def test_calculate_methane_emissions(vol_sols: float, temp: float, degradable: bool, expected: float) -> None:
     """Test that methane emissions from a storage are calculated correctly."""
@@ -53,33 +51,3 @@ def test_calculate_arrhenius_exponent_error(temp: float) -> None:
     """Test that Arrhenius exponent equation raises an error when passed an invalid temperature."""
     with pytest.raises(ValueError):
         Storage._calculate_arrhenius_exponent(temp)
-
-
-@pytest.mark.parametrize(
-    "total_ammoniacal, volume, density, temp, area, pH, expected",
-    [
-        (0.0, 1_000.0, 10.0, 20.0, 1_500.0, 8.0, 0.0),
-        (25.0, 800.0, 15.0, 5.0, 300.0, 6.5, 25.0)
-    ]
-)
-def test_calculate_ammonia_emissions(mocker: MockerFixture, total_ammoniacal: float, volume: float, density: float, temp: float, area: float, pH: float, expected: float) -> None:
-    """Test that ammonia emissions from a storage are calculated correctly."""
-    mocker.patch.object(Storage, "_calculate_ammonia_equilibrium_coefficient", return_value=0.1)
-
-    actual = Storage._calculate_ammonia_emissions(total_ammoniacal, volume, density, temp, area, pH)
-
-    assert pytest.approx(actual) == expected
-
-
-@pytest.mark.parametrize(
-    "total_ammoniacal, volume, density, area", [
-        (-1_000.0, 3_000.0, 44.0, 500.0),
-        (1_000.0, -3_000.0, 44.0, 500.0),
-        (1_000.0, 3_000.0, -44.0, 500.0),
-        (1_000.0, 3_000.0, 44.0, -500.0)
-    ]
-)
-def test_calculate_ammonia_emissions_error(total_ammoniacal: float, volume: float, density: float, area: float) -> None:
-    """Test that ammonia emissions calculations raise an error when passed an invalid value."""
-    with pytest.raises(ValueError):
-        Storage._calculate_ammonia_emissions(total_ammoniacal, volume, density, 20.0, area, 6.0)
